server/requirements.txt
Django >= 3.1.5, <3.2.0
djangorestframework >=3.12.2, <3.14.0
flake8 >= 3.8.4, <3.9.0
psycopg2 >=2.8.6, <2.9.0
drf-yasg == 1.20.0